adding changes to get balance

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -5,7 +5,7 @@ import com.techelevator.tenmo.model.Account;
 import java.math.BigDecimal;
 
 public interface AccountDao {
-    BigDecimal getBalance(int userid);
+    Integer getBalance(int userid);
